requirements.txt
python-telegram-bot[rate-limiter]==21.4
fastapi==0.112.2
uvicorn==0.30.6
aiosqlite==0.20.0
pytz==2024.2
python-dotenv==1.0.1